requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2
EbookLib==0.17.1
idna==3.3
lxml==4.7.1
Pillow==9.0.0
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
